♻️ refactor: footer.tsx を Footer.tsx にリネーム (#85)
他のコンポーネントファイルに合わせて PascalCase に統一。
import パスを参照する App.tsx, Layout.tsx も合わせて修正。
